requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.12.5
mysql-connector-python==8.2.0
python-dotenv==1.2.1